Change DTO and services

+++ Lab6-7/Lab6-7.BLL/Services/Contractor/ContractorService.cs
@@ -3,6 +3,7 @@ using Lab6_7.BLL.DTOs.Contractor;
 using Lab6_7.BLL.Interfaces.Contractor;
 using Lab6_7.DataAccess.Models.Contractor;
 using Lab6_7.DataAccess.Repositories.Interfaces.Contractor;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -25,6 +26,11 @@ namespace Lab6_7.BLL.Services.Contractor
         {
             var contractor = _mapper.Map<ContractorDTO, ContractorModel>(contractorDTO);
 
+            if(contractor == null)
+            {
+                throw new ArgumentException("Exception just for fun");
+            }
+
             await _contractorRepo.CreateAsync(contractor);
 
             await _contractorRepo.SaveAsync();
@@ -33,6 +39,11 @@ namespace Lab6_7.BLL.Services.Contractor
         public async Task ChangeContractorAsync(ContractorDTO contractorDTO)
         {
             var contractor = await _contractorRepo.GetFirstOrDefaultAsync(c => c.Id == contractorDTO.Id);
+
+            if (contractor == null)
+            {
+                throw new ArgumentNullException("Contractor doesn't exist");
+            }
 
             _contractorRepo.Update(contractor);
 
@@ -43,6 +54,11 @@ namespace Lab6_7.BLL.Services.Contractor
         {
             var contractor = await _contractorRepo.GetFirstOrDefaultAsync(c => c.Id == contractorDTO.Id);
 
+            if (contractor == null)
+            {
+                throw new ArgumentNullException("Contractor doesn't exist");
+            }
+
             _contractorRepo.Delete(contractor);
 
             await _contractorRepo.SaveAsync();
@@ -51,6 +67,11 @@ namespace Lab6_7.BLL.Services.Contractor
         public async Task DeleteContractorAsync(int id)
         {
             var contractor = await _contractorRepo.GetFirstOrDefaultAsync(c => c.Id == id);
+
+            if (contractor == null)
+            {
+                throw new ArgumentNullException("Contractor doesn't exist");
+            }
 
             _contractorRepo.Delete(contractor);
